Refactor ProtectedRoute component into its own file
